ML-backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
eventlet==0.39.1
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.56.0
fsspec==2025.3.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.30.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pymongo==4.11.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
simple-websocket==1.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.86
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
